0.3.4: Vastly improve QR behavior
No more endless printings of QR codes. It seems that forcing the
module_color and what not makes it work on normal and reverse
terminals. Yay!

Now it prints the QR code once. If it's not visible, resize and
try again.

I might be able to add row detection and say that the terminal is
not wide enough. Hmm.